Enhancements related to SFDX project initialization from an existing production org (#96)
* New cleaning commands + deploy tips

* new parameter websocket

* Indicating in logs when deployment is a simulation

* Build & lint fixes

* Enhance scratch create logs + lint fix

* Fix sfdmu call

* [Mega-Linter] Apply linters fixes

+++ src/commands/hardis/project/create.ts
@@ -29,6 +29,9 @@ export default class ProjectCreate extends SfdxCommand {
       default: false,
       description: messages.getMessage("debugMode"),
     }),
+    websocket: flags.string({
+      description: messages.getMessage("websocket"),
+    }),
   };
 
   // Comment this out if your command does not require an org username
@@ -65,22 +68,24 @@ export default class ProjectCreate extends SfdxCommand {
         name: "projectName",
         message: "What is the name of your project ?",
       });
-      // Create sfdx project
       projectName = projectRes.projectName.toLowerCase().replace(" ", "_");
     }
-    const createCommand = "sfdx force:project:create" + ` --projectname "${projectName}"` + " --manifest";
-    await execCommand(createCommand, this, {
-      output: true,
-      fail: true,
-      debug: this.debugMode,
-    });
 
-    // Move project files at root
-    await fs.copy(path.join(process.cwd(), projectName), process.cwd(), {
-      overwrite: false,
-    });
-    await fs.rmdir(path.join(process.cwd(), projectName), { recursive: true });
+    // Create sfdx project only if not existing
+    if (!fs.existsSync("./sfdx-project.json")) {
+      const createCommand = "sfdx force:project:create" + ` --projectname "${projectName}"` + " --manifest";
+      await execCommand(createCommand, this, {
+        output: true,
+        fail: true,
+        debug: this.debugMode,
+      });
 
+      // Move project files at root
+      await fs.copy(path.join(process.cwd(), projectName), process.cwd(), {
+        overwrite: false,
+      });
+      await fs.rmdir(path.join(process.cwd(), projectName), { recursive: true });
+    }
     // Copy default project files
     uxLog(this, "Copying default files...");
     await fs.copy(path.join(__dirname, "../../../../defaults/ci", "."), process.cwd(), { overwrite: false });
